Relocate the file and add notes for instructor
Relocated the file to the _resources folder so that it shows up on the
'resources' page under the 'Building' heading. I figured that was a
better spot than the 'curriculum' page.

I also added the instructor notes to the bottom of the page, that were
initially in the original commit message.